Move migrations to a new file
